+++ android/app/src/main/java/com/bookease/app/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.anonymous.BookEaseApp
+package com.bookease.app
 import expo.modules.splashscreen.SplashScreenManager
 
 import android.os.Build