+++ metro.config.js
@@ -1,4 +1,3 @@
-// Learn more https://docs.expo.io/guides/customizing-metro
 const { getDefaultConfig } = require('expo/metro-config');
 
 /** @type {import('expo/metro-config').MetroConfig} */
@@ -6,5 +5,11 @@ const config = getDefaultConfig(__dirname, {
     // [Web-only] Enables CSS support in Metro.
     isCSSEnabled: true,
 });
+
+config.resolver.blockList = [
+    ...(config.resolver.blockList || []),
+    /\/electron\/.*/,
+    /\/dist-electron\/.*/,
+];
 
 module.exports = config;